requirements.txt
apache-airflow==2.7.3
yfinance==0.2.33
pandas==2.1.3
pyarrow==14.0.1
boto3==1.33.6
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
requests==2.31.0
